Removing ImageAsset adding ImageObject3D
added image editor
using image editor for public property editor
added ring object
