--- DesignTools/ImageAsset.cs
@@ -1,91 +1,0 @@
-﻿/*
-Copyright (c) 2018, Lars Brubaker, John Lewin
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-
-1. Redistributions of source code must retain the above copyright notice, this
-   list of conditions and the following disclaimer.
-2. Redistributions in binary form must reproduce the above copyright notice,
-   this list of conditions and the following disclaimer in the documentation
-   and/or other materials provided with the distribution.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-The views and conclusions contained in the software and documentation are those
-of the authors and should not be interpreted as representing official policies,
-either expressed or implied, of the FreeBSD Project.
-*/
-
-using System.IO;
-using MatterHackers.Agg.Image;
-using MatterHackers.Agg.ImageProcessing;
-using MatterHackers.Agg.Platform;
-using Newtonsoft.Json;
-
-namespace MatterHackers.MatterControl.DesignTools
-{
-	public class ImageAsset
-	{
-		private string _assetPath;
-		private ImageBuffer _image;
-		private bool _invert;
-
-		public string AssetPath
-		{
-			get { return _assetPath; }
-			set
-			{
-				if (_assetPath != value)
-				{
-					_assetPath = value;
-					_image = null;
-				}
-			}
-		}
-
-		[JsonIgnore]
-		public ImageBuffer Image
-		{
-			get
-			{
-				if (_image == null)
-				{
-					if (File.Exists(AssetPath))
-					{
-						_image = AggContext.ImageIO.LoadImage(AssetPath);
-						if (Invert)
-						{
-							_image = InvertLightness.DoInvertLightness(_image);
-						}
-					}
-				}
-
-				return _image;
-			}
-		}
-
-		public bool Invert
-		{
-			get { return _invert; }
-			set
-			{
-				if (_invert != value)
-				{
-					_invert = value;
-					_image = null;
-				}
-			}
-		}
-	}
-}
